requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
celery
facebook-business
google-ads
pydantic
python-dotenv
openai
httpx
reportlab
PyJWT
passlib[bcrypt]
